Refactor: 전역 인증 상태 관리 구조 개선
